requirements.txt
streamlit==1.31.0
ultralytics==8.1.0
opencv-python==4.9.0.80
pillow==10.2.0
google-generativeai==0.3.2
numpy==1.26.4
deep-translator==1.11.4
SpeechRecognition==3.10.1
pyttsx3==2.90
torch==2.2.0
torchvision==0.17.0
watchdog==3.0.0
folium==0.12.1
streamlit-folium==0.7.0
geopy==2.2.0
requests==2.26.0 